fields関連の表示をbootstrapに

+++ src/Controller/FieldsController.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace App\Controller;
 
 use App\Controller\AppController;
@@ -10,23 +11,29 @@ use App\Controller\AppController;
  *
  * @method \App\Model\Entity\Field[] paginate($object = null, array $settings = [])
  */
-class FieldsController extends AppController
-{
+class FieldsController extends AppController {
 
     /**
      * Index method
      *
      * @return \Cake\Http\Response|void
      */
-    public function index()
-    {
+    public function index() {
         $this->paginate = [
-            'contain' => ['Organizations', 'ParentFields']
+            'contain' => ['Organizations', 'ParentFields'],
+            'order' => ['lft' => 'ASC',]
         ];
-        $fields = $this->paginate($this->Fields);
+        
+        $query =$this->Fields->find('depth')
+                ->select( $this->Fields )
+                ->select( $this->Fields->Organizations )
+                ;
+        
+        $fields = $this->paginate($query);
 
         $this->set(compact('fields'));
         $this->set('_serialize', ['fields']);
+        $this->viewBuilder()->layout('bootstrap');
     }
 
     /**
@@ -36,8 +43,7 @@ class FieldsController extends AppController
      * @return \Cake\Http\Response|void
      * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
      */
-    public function view($id = null)
-    {
+    public function view($id = null) {
         $field = $this->Fields->get($id, [
             'contain' => ['Organizations', 'ParentFields', 'ChildFields', 'Skills']
         ]);
@@ -51,8 +57,7 @@ class FieldsController extends AppController
      *
      * @return \Cake\Http\Response|null Redirects on successful add, renders view otherwise.
      */
-    public function add()
-    {
+    public function add() {
         $field = $this->Fields->newEntity();
         if ($this->request->is('post')) {
             $field = $this->Fields->patchEntity($field, $this->request->getData());
@@ -63,10 +68,15 @@ class FieldsController extends AppController
             }
             $this->Flash->error(__('The field could not be saved. Please, try again.'));
         }
+
         $organizations = $this->Fields->Organizations->find('list', ['limit' => 200]);
         $parentFields = $this->Fields->ParentFields->find('list', ['limit' => 200]);
+        $parentFields->where([$this->Fields->ParentFields->aliasField('id') . ' is not' => $field->id]);
+
         $this->set(compact('field', 'organizations', 'parentFields'));
         $this->set('_serialize', ['field']);
+        $this->viewBuilder()->layout('bootstrap');
+        $this->render('edit');
     }
 
     /**
@@ -76,8 +86,7 @@ class FieldsController extends AppController
      * @return \Cake\Http\Response|null Redirects on successful edit, renders view otherwise.
      * @throws \Cake\Network\Exception\NotFoundException When record not found.
      */
-    public function edit($id = null)
-    {
+    public function edit($id = null) {
         $field = $this->Fields->get($id, [
             'contain' => []
         ]);
@@ -92,8 +101,11 @@ class FieldsController extends AppController
         }
         $organizations = $this->Fields->Organizations->find('list', ['limit' => 200]);
         $parentFields = $this->Fields->ParentFields->find('list', ['limit' => 200]);
+        $parentFields->where([$this->Fields->ParentFields->aliasField('id') . ' is not' => $field->id]);
+
         $this->set(compact('field', 'organizations', 'parentFields'));
         $this->set('_serialize', ['field']);
+        $this->viewBuilder()->layout('bootstrap');
     }
 
     /**
@@ -103,10 +115,16 @@ class FieldsController extends AppController
      * @return \Cake\Http\Response|null Redirects to index.
      * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
      */
-    public function delete($id = null)
-    {
+    public function delete($id = null) {
         $this->request->allowMethod(['post', 'delete']);
+        
         $field = $this->Fields->get($id);
+        
+        if ($field->rght - $field->lft > 1) {
+            $this->Flash->error('下位グループを持つグループは削除できません');
+            return $this->redirect(['action' => 'index']);
+        }
+
         if ($this->Fields->delete($field)) {
             $this->Flash->success(__('The field has been deleted.'));
         } else {
@@ -115,4 +133,5 @@ class FieldsController extends AppController
 
         return $this->redirect(['action' => 'index']);
     }
+
 }
