+++ src/Controller/OrganizationsController.php
@@ -1,7 +1,10 @@
 <?php
+
 namespace App\Controller;
 
 use App\Controller\AppController;
+use App\Defines\Defines;
+use Cake\Utility\Hash;
 
 /**
  * Organizations Controller
@@ -10,23 +13,37 @@ use App\Controller\AppController;
  *
  * @method \App\Model\Entity\Organization[] paginate($object = null, array $settings = [])
  */
-class OrganizationsController extends AppController
-{
+class OrganizationsController extends AppController {
+
+    public $helpers = [
+        'Paginator' => ['templates' => 'paginator-templates']
+    ];
 
     /**
      * Index method
      *
      * @return \Cake\Http\Response|void
      */
-    public function index()
-    {
+    public function index() {
+        $user = $this->Auth->user();
+        $group = Hash::get($user, 'group_id', 0);
+
         $this->paginate = [
-            'contain' => ['ParentOrganizations']
+            'contain' => ['ParentOrganizations'],
+            'order' => ['lft' => 'ASC',]
         ];
-        $organizations = $this->paginate($this->Organizations);
+
+        $query = $this->Organizations->find('depth');
+
+        if ($group == Defines::GROUP_ORGANIZATION_ADMIN) {
+            $query->find('user', ['user_id' => $user->id, 'relation' => 'children']);
+        }
+
+        $organizations = $this->paginate($query);
 
         $this->set(compact('organizations'));
         $this->set('_serialize', ['organizations']);
+        $this->viewBuilder()->layout('bootstrap');
     }
 
     /**
@@ -36,8 +53,7 @@ class OrganizationsController extends AppController
      * @return \Cake\Http\Response|void
      * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
      */
-    public function view($id = null)
-    {
+    public function view($id = null) {
         $organization = $this->Organizations->get($id, [
             'contain' => ['ParentOrganizations', 'Users', 'Fields', 'ChildOrganizations', 'OrganizationsUsers']
         ]);
@@ -51,8 +67,10 @@ class OrganizationsController extends AppController
      *
      * @return \Cake\Http\Response|null Redirects on successful add, renders view otherwise.
      */
-    public function add()
-    {
+    public function add() {
+        $user = $this->Auth->user();
+        $group = Hash::get($user, 'group_id', 0);
+
         $organization = $this->Organizations->newEntity();
         if ($this->request->is('post')) {
             $organization = $this->Organizations->patchEntity($organization, $this->request->getData());
@@ -63,10 +81,17 @@ class OrganizationsController extends AppController
             }
             $this->Flash->error(__('The organization could not be saved. Please, try again.'));
         }
+
         $parentOrganizations = $this->Organizations->ParentOrganizations->find('list', ['limit' => 200]);
-        $users = $this->Organizations->Users->find('list', ['limit' => 200]);
-        $this->set(compact('organization', 'parentOrganizations', 'users'));
+
+        if ($group == Defines::GROUP_ORGANIZATION_ADMIN) {
+            $parentOrganizations->find('user', ['user_id' => $user->id, 'relation' => 'children']);
+        }
+
+        $this->set(compact('organization', 'parentOrganizations'));
         $this->set('_serialize', ['organization']);
+        $this->viewBuilder()->layout('bootstrap');
+        $this->render('edit');
     }
 
     /**
@@ -76,10 +101,12 @@ class OrganizationsController extends AppController
      * @return \Cake\Http\Response|null Redirects on successful edit, renders view otherwise.
      * @throws \Cake\Network\Exception\NotFoundException When record not found.
      */
-    public function edit($id = null)
-    {
+    public function edit($id = null) {
+        $user = $this->Auth->user();
+        $group = Hash::get($user, 'group_id', 0);
+
+
         $organization = $this->Organizations->get($id, [
-            'contain' => ['Users']
         ]);
         if ($this->request->is(['patch', 'post', 'put'])) {
             $organization = $this->Organizations->patchEntity($organization, $this->request->getData());
@@ -90,10 +117,19 @@ class OrganizationsController extends AppController
             }
             $this->Flash->error(__('The organization could not be saved. Please, try again.'));
         }
+
         $parentOrganizations = $this->Organizations->ParentOrganizations->find('list', ['limit' => 200]);
+        $parentOrganizations->where([$this->Organizations->ParentOrganizations->aliasField('id') . ' is not' => $organization->id]);
+
+        if ($group == Defines::GROUP_ORGANIZATION_ADMIN) {
+            $parentOrganizations->find('user', ['user_id' => $user->id, 'relation' => 'children']);
+        }
+
+
         $users = $this->Organizations->Users->find('list', ['limit' => 200]);
         $this->set(compact('organization', 'parentOrganizations', 'users'));
         $this->set('_serialize', ['organization']);
+        $this->viewBuilder()->layout('bootstrap');
     }
 
     /**
@@ -103,10 +139,15 @@ class OrganizationsController extends AppController
      * @return \Cake\Http\Response|null Redirects to index.
      * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
      */
-    public function delete($id = null)
-    {
+    public function delete($id = null) {
         $this->request->allowMethod(['post', 'delete']);
         $organization = $this->Organizations->get($id);
+
+        if ($organization->rght - $organization->lft > 1) {
+            $this->Flash->error('下位組織を持つ組織は削除できません');
+            return $this->redirect(['action' => 'index']);
+        }
+
         if ($this->Organizations->delete($organization)) {
             $this->Flash->success(__('The organization has been deleted.'));
         } else {
@@ -115,4 +156,5 @@ class OrganizationsController extends AppController
 
         return $this->redirect(['action' => 'index']);
     }
+
 }
